apps/discord/requirements.txt
discord.py
python-dotenv
firebase-admin